feat: 初步实现公开信箱
* 还有很多杂七杂八的

+++ app/utils/dateFormat.ts
@@ -1,4 +1,4 @@
 export const dateFormat = new Intl.DateTimeFormat("zh-CN", {
   dateStyle: "long",
-  timeStyle: "short",
+  timeStyle: "long",
 });
